chore: organizando o projeto
