+++ src/constants/status.ts
@@ -1,0 +1,11 @@
+const STATUS = {
+  OK: 200,
+  CREATED: 201,
+  UPDATED: 201,
+  NOT_FOUND: 400,
+  BAD_REQUEST: 400,
+  INTERNAL_SERVER_ERROR: 500,
+  DEFAULT_ERROR: 418,
+};
+
+export default STATUS;
