+++ src/tests/all.spec.ts
@@ -6,10 +6,10 @@ import * as sinon from "sinon";
 import { faker } from "@faker-js/faker";
 import { expect, assert } from "chai";
 
-import "./database";
-import { Region, RegionModel, UserModel } from "./models";
-import GeoLib from "./lib";
-import server from "./server";
+import "../database/database";
+import { Region, RegionModel, UserModel } from "../models/models";
+import GeoLib from "../libs/geolib";
+import server from "../server";
 
 describe("Models", () => {
   let user;
